refs #20074: refactoring
